+++ extensions/reclaim-ai/src/hooks/useSchedulingLinks.tsx
@@ -1,42 +1,21 @@
-import { getPreferenceValues, showToast, Toast, open } from "@raycast/api";
-import { useFetch } from "@raycast/utils";
-import { useMemo } from "react";
-import { NativePreferences } from "../types/preferences";
-import { ApiSchedulingLink, ApiSchedulingLinkGroups } from "./useSchedulingLinks.types";
+import { open, showToast, Toast } from "@raycast/api";
 import { SchedulingLink } from "../types/scheduling-link";
 import useApi from "./useApi";
+import { ApiSchedulingLink, ApiSchedulingLinkGroups } from "./useSchedulingLinks.types";
+import { fetchPromise } from "../utils/fetcher";
 
 export const useSchedulingLinks = () => {
-  const { apiUrl, apiToken } = getPreferenceValues<NativePreferences>();
-
-  const headers = useMemo(
-    () => ({
-      Authorization: `Bearer ${apiToken}`,
-      "Content-Type": "application/json",
-      Accept: "application/json",
-    }),
-    [apiToken]
-  );
-
   const {
     data: schedulingLinks,
     error: schedulingLinksError,
     isLoading: schedulingLinksIsLoading,
-  } = useFetch<ApiSchedulingLink>(`${apiUrl}/scheduling-link`, {
-    headers,
-    keepPreviousData: true,
-    method: "GET",
-  });
+  } = useApi<ApiSchedulingLink>("/scheduling-link");
 
   const {
     data: schedulingLinksGroups,
     error: schedulingLinksGroupsError,
     isLoading: schedulingLinksGroupsIsLoading,
-  } = useFetch<ApiSchedulingLinkGroups>(`${apiUrl}/scheduling-link/group`, {
-    headers,
-    keepPreviousData: true,
-    method: "GET",
-  });
+  } = useApi<ApiSchedulingLinkGroups>("/scheduling-link/group");
 
   if (schedulingLinksError) console.error("Error while fetching Scheduling Links", schedulingLinksError);
   if (schedulingLinksGroupsError)
@@ -54,8 +33,6 @@ export const useSchedulingLinks = () => {
 
 export const useSchedulingLinkActions = (link: SchedulingLink) => {
   const createOneOffLink = async () => {
-    const { fetchPromise } = useApi();
-
     const [oneOff, error] = await fetchPromise<SchedulingLink>(
       "/scheduling-link/derivative",
       { method: "POST" },