+++ extensions/reclaim-ai/src/hooks/useTask.tsx
@@ -1,32 +1,12 @@
-import { getPreferenceValues, showHUD } from "@raycast/api";
-import { useFetch } from "@raycast/utils";
-import { useMemo } from "react";
-import { NativePreferences } from "../types/preferences";
+import { showHUD } from "@raycast/api";
+import { RequestInit } from "node-fetch";
 import { Task } from "../types/task";
 import useApi from "./useApi";
 import { CreateTaskProps, PlannerActionIntermediateResult } from "./useTask.types";
-import { RequestInit } from "node-fetch";
+import { fetchPromise } from "../utils/fetcher";
 
 export const useTasks = () => {
-  const { apiUrl, apiToken } = getPreferenceValues<NativePreferences>();
-
-  const headers = useMemo(
-    () => ({
-      Authorization: `Bearer ${apiToken}`,
-      "Content-Type": "application/json",
-      Accept: "application/json",
-    }),
-    [apiToken]
-  );
-
-  const {
-    data: tasks,
-    error,
-    isLoading,
-  } = useFetch<Task[]>(`${apiUrl}/tasks?instances=true`, {
-    headers,
-    keepPreviousData: true,
-  });
+  const { data: tasks, error, isLoading } = useApi<Task[]>("/tasks?instances=true");
 
   if (error) console.error("Error while fetching Tasks", error);
 
@@ -38,8 +18,6 @@ export const useTasks = () => {
 };
 
 export const useTaskActions = () => {
-  const { fetchPromise } = useApi();
-
   const executeTaskAction = async <T,>(url: string, options?: RequestInit, payload?: unknown): Promise<T> => {
     const [response, error] = await fetchPromise<T>(url, options, payload);
     if (error) throw error;